% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_analysis}
\alias{app_analysis}
\title{app_analysis}
\usage{
app_analysis(input, output, session)
}
\arguments{
\item{input}{UI input}

\item{output}{server output}

\item{session}{number of session}
}
\description{
app_analysis
}
