% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_map}
\alias{app_map}
\title{app_map}
\usage{
app_map(input, output, session)
}
\arguments{
\item{input}{UI input}

\item{output}{server output}

\item{session}{number of session}
}
\value{
map
}
\description{
app_map
}
