% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Main server function app_server}
\usage{
app_server(input, output)
}
\arguments{
\item{input}{UI input}

\item{output}{server output}
}
\value{
application
}
\description{
Main server function app_server
}
