% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departement.R
\docType{data}
\name{departement}
\alias{departement}
\title{French departements and regions file}
\format{A data frame with 5032 rows and 34 variables:
\describe{
  \item{CODE_DEPT}{department code}
  \item{NOM_REG}{region name}
  \item{geometry}{shape of the department France in crs 2154}
  }}
\usage{
departement
}
\description{
A dataset containing shapefile for France's departments and region.
}
\keyword{datasets}
