% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_cote.R
\name{fonction_cote}
\alias{fonction_cote}
\title{fonction_cote}
\usage{
fonction_cote(gareA, gareD, mois)
}
\arguments{
\item{mois}{}
}
\value{
the odds of a train to be late given its departure and arrival stations and the month of the train
}
\description{
fonction_cote
}
\examples{
fonction_cote("PARIS MONTPARNASSE", "NANTES", 12)
}
