% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_temps_moyen.R
\name{fonction_temps_moyen}
\alias{fonction_temps_moyen}
\title{fonction_temps_moyen function}
\usage{
fonction_temps_moyen(gareA, gareB)
}
\arguments{
\item{gareB}{}
}
\value{
average delay
}
\description{
fonction_temps_moyen function
}
\examples{
fonction_temps_moyen(gareA = "METZ", gareB = "PARIS EST")
}
