% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocoding}
\alias{geocoding}
\title{Geocoding}
\usage{
geocoding(address = NULL)
}
\arguments{
\item{address}{is the address of the railway staytion. There is no special formatting, but the preferred one is: "gare de ..."}
}
\value{
data frame with latitude and longitude of said "google map" request
}
\description{
Geocoding
}
\examples{
geocoding("nantes, gare")
geocoding("gare de nantes")$lon
geocoding("gare, nantes")$lat
}
