% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNCF_regularite.R
\docType{data}
\name{SNCF_regularite}
\alias{SNCF_regularite}
\title{Information about the SNCF trains starting from 2015.}
\format{A data frame with 5032 rows and 34 variables:
\describe{
  \item{annee}{year}
  \item{mois}{month}
  ...
}}
\source{
\url{https://data.sncf.com/explore/dataset/regularite-mensuelle-tgv-aqst/table/?sort=periode&fbclid=IwAR1iEGCOzJKKqUCVA-zJLW8-3w9zKi1R7LhYYd5BsBVlqvCCZvyafZlUqN4}
}
\usage{
SNCF_regularite
}
\description{
A dataset containing the statistics and other information about the delays of SNCF trains.
}
\keyword{datasets}
